Use trait for custom media properties
