+++ src/Models/Traits/CustomMediaProperties.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Spatie\MediaLibrary\Models\Traits;
+
+trait CustomMediaProperties
+{
+    public function setCustomHeaders(array $customHeaders): self
+    {
+        $this->setCustomProperty('custom_headers', $customHeaders);
+
+        return $this;
+    }
+
+    public function getCustomHeaders(): array
+    {
+        return $this->getCustomProperty('custom_headers', []);
+    }
+}
